+++ src/app/discovery.ts
@@ -1,10 +1,9 @@
 export interface Service {
-  path?: string
-  hostname?: string
-  id?: string
-  address: string
-  port?: number
-  name: string,
-  icon?: string // Url to an icon
+  path?: string;
+  hostname?: string;
+  id?: string;
+  address: string;
+  port?: number;
+  name: string;
+  icon?: string; // Url to an icon
 }
-