+++ src/app/settings.service.ts
@@ -1,24 +1,26 @@
 import { Injectable } from '@angular/core';
 import { ActionSheetController } from '@ionic/angular';
 import { Service } from './discovery';
+import { App } from '@capacitor/app';
 
 export enum Role {
   destructive = 'destructive',
   cancel = 'cancel',
   go = 'go',
-  privacy = 'privacy'
+  privacy = 'privacy',
 }
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class SettingsService {
-
-  constructor() { }
+  constructor() {}
 
   public async presentSettings(ctrl: ActionSheetController): Promise<Role> {
+    const info = await App.getInfo();
     const actionSheet = await ctrl.create({
       header: 'Settings',
+      subHeader: `v${info.version}.${info.build}`,
       buttons: [
         { text: 'Privacy Policy', role: Role.privacy },
         {
@@ -46,7 +48,6 @@ export class SettingsService {
   }
 
   public async presentActions(ctrl: ActionSheetController, service: Service): Promise<Role> {
-
     const delBtn = {
       text: 'Delete',
       role: Role.destructive,
@@ -64,18 +65,21 @@ export class SettingsService {
     let buttons = [delBtn, cancelBtn];
 
     if (service.port) {
-      buttons = [{
-        text: service.address + ':' + service.port,
-        role: Role.go,
-        data: {
-          action: 'go',
+      buttons = [
+        {
+          text: service.address + ':' + service.port,
+          role: Role.go,
+          data: {
+            action: 'go',
+          },
         },
-      }, cancelBtn];
+        cancelBtn,
+      ];
     }
     const actionSheet = await ctrl.create({
       header: service.name,
       subHeader: service.hostname,
-      buttons
+      buttons,
     });
 
     await actionSheet.present();