+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## 1.3
+- Hide QR option on web
+- Fix removal of old detected shortcuts
+- Use larger Capacitor icon
+- Fix when users add www. at front of urls
+- Allow close of privacy page
+
 ## 1.2
 - Add support for appflow.app.com QR codes
+- Associate Domain nexusbrowser.com to app